Cache npm bin paths.
The eslint path is set for each JS-file, and calling 'npm bin' is quite
expensive (>.3sec). This is distracting during file opening, and particularly
slow when a previous session is started that has many tabs/windows open.

Also, keep in mind that projects can be highly modularied with many subdirectories.

Therefore, cache the eslint-paths per project:

The assumption is that eslint is found in <project-dir>/node_modules/.bin
Opening the first JS-file will cache the eslint-path keyed with <project-dir>.
Opening subsequent files which are located somewhere below <project-dir> will use
the cached eslint-path.